tests/wpt/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.roundrect.2.radii.2.dompoint.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.path.roundrect.2.radii.2.dompoint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.2.radii.2.dompoint</h1>
<p class="desc">Verify that when two radii are given to roundRect(), the second radius, specified as a DOMPoint, applies to the top-right and bottom-left corners.</p>


<script>
var t = async_test("Verify that when two radii are given to roundRect(), the second radius, specified as a DOMPoint, applies to the top-right and bottom-left corners.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.roundRect(0, 0, 100, 50, [0, new DOMPoint(40, 20)]);
ctx.fillStyle = '#0f0';
ctx.fill();

// top-right corner
_assertPixel(offscreenCanvas, 79,1, 255,0,0,255, "79,1", "255,0,0,255");
_assertPixel(offscreenCanvas, 58,1, 0,255,0,255, "58,1", "0,255,0,255");
_assertPixel(offscreenCanvas, 98,10, 255,0,0,255, "98,10", "255,0,0,255");
_assertPixel(offscreenCanvas, 98,21, 0,255,0,255, "98,21", "0,255,0,255");

// bottom-left corner
_assertPixel(offscreenCanvas, 20,48, 255,0,0,255, "20,48", "255,0,0,255");
_assertPixel(offscreenCanvas, 41,48, 0,255,0,255, "41,48", "0,255,0,255");
_assertPixel(offscreenCanvas, 1,39, 255,0,0,255, "1,39", "255,0,0,255");
_assertPixel(offscreenCanvas, 1,28, 0,255,0,255, "1,28", "0,255,0,255");

// other corners
_assertPixel(offscreenCanvas, 1,1, 0,255,0,255, "1,1", "0,255,0,255");
_assertPixel(offscreenCanvas, 98,48, 0,255,0,255, "98,48", "0,255,0,255");
t.done();

});
</script>
